define: readline-interface.jsを作成、import readline from 'readline'に変更することでよりわかりやすく
